Phase 5: extract admin merge form JS to external file
Move GameMergeFormController and streaming merge progress handling from
admin/merge.php into admin-merge-form.js and load it with StaticAsset
cache-busting.
